Updated copyright, new game engine module
The game engine module will be where I move all the global variables,
and will help me organize the structure of the whole project much
better. Along with that, I hope to eliminate any global variables being
used inside of other modules. Not only will this help with the code
structure, but it will most likely be nessecary if I am to attempt
making this game network portable (multiplayer).